requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
openai
google-generativeai
nltk
python-multipart
requests
